refactor: 重构项目结构与依赖配置 1. 调整Gradle插件使用方式为libs插件引用，清理冗余gradle配置 2. 重构UI页面包结构，将所有Activity迁移至ui.activity包 3. ViewModel命名修改
