+++ app/src/main/AndroidManifest.xml
@@ -47,14 +47,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.NotifyRelay">
-        <activity android:name=".ui.GuideActivity"
+        <activity android:name=".ui.activity.GuideActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.MainActivity" android:exported="true" />
+        <activity android:name=".ui.activity.MainActivity" android:exported="true" />
         <!-- 通知监听服务声明，用于系统识别通知访问权限 -->
         <service
             android:name=".servers.NotifyRelayNotificationListenerService"
@@ -101,7 +101,7 @@
             
         <!-- 剪贴板同步透明Activity -->
         <activity
-            android:name=".ui.ClipboardSyncActivity"
+            android:name=".ui.activity.ClipboardSyncActivity"
             android:exported="true"
             android:theme="@style/TransparentTheme"
             android:launchMode="singleInstance"
@@ -111,7 +111,7 @@
             
         <!-- 开发者模式Activity -->
         <activity
-            android:name=".ui.DeveloperModeActivity"
+            android:name=".ui.activity.DeveloperModeActivity"
             android:exported="true"
             android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar" />
