+++ app/src/main/java/com/xzyht/notifyrelay/ui/viewmodel/SuperIslandHistoryViewModel.kt
@@ -1,4 +1,4 @@
-package com.xzyht.notifyrelay.ui.ViewModels
+package com.xzyht.notifyrelay.ui.viewmodel
 
 import android.app.Application
 import android.graphics.Bitmap
@@ -9,8 +9,8 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
-import com.xzyht.notifyrelay.feature.device.model.NotificationRepository
-import com.xzyht.notifyrelay.feature.notification.backend.RemoteFilterConfig
+import com.xzyht.notifyrelay.feature.notification.superisland.history.SuperIslandHistoryStore
+import com.xzyht.notifyrelay.feature.notification.superisland.history.SuperIslandHistoryStoreEntry
 import com.xzyht.notifyrelay.servers.appslist.AppRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -18,64 +18,33 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import notifyrelay.data.database.repository.DatabaseRepository
 
-class NotificationHistoryViewModel(
+class SuperIslandHistoryViewModel(
     private val application: Application,
     private val repository: DatabaseRepository
 ) : ViewModel() {
-    private val _uiState = MutableStateFlow(NotificationHistoryUiState())
-    val uiState: StateFlow<NotificationHistoryUiState> = _uiState.asStateFlow()
+    private val _uiState = MutableStateFlow(SuperIslandHistoryUiState())
+    val uiState: StateFlow<SuperIslandHistoryUiState> = _uiState.asStateFlow()
 
     private val _appIconCache = MutableStateFlow<Map<String, Pair<String, Bitmap?>>>(emptyMap())
-    val appIconCache: StateFlow<Map<String, Pair<String, Bitmap?>>> = _appIconCache.asStateFlow()
 
-    private val installedPackages = MutableStateFlow<Set<String>>(emptySet())
-    val installedPackagesState: StateFlow<Set<String>> = installedPackages.asStateFlow()
-
-    private val deviceFlow = MutableStateFlow("本机")
     private val refreshSignal = MutableStateFlow(0L)
     private val iconLoading = mutableSetOf<String>()
 
     @OptIn(ExperimentalCoroutinesApi::class)
-    val groupedPagingFlow: Flow<PagingData<GroupedNotifications>> = combine(
-        deviceFlow,
-        refreshSignal,
-        installedPackages
-    ) { device, _, packages ->
-        device to packages
-    }.flatMapLatest { (device, packages) ->
+    val groupedPagingFlow: Flow<PagingData<GroupedSuperIslandHistory>> = refreshSignal.flatMapLatest {
         Pager(
             config = PagingConfig(pageSize = 20, enablePlaceholders = false)
         ) {
-            NotificationPagingSource(
-                repository = repository,
-                deviceUuid = device,
-                installedPackages = packages
-            )
+            SuperIslandPagingSource(repository, application)
         }.flow
     }.cachedIn(viewModelScope)
 
     init {
-        viewModelScope.launch(Dispatchers.IO) {
-            if (!RemoteFilterConfig.isLoaded) {
-                RemoteFilterConfig.load(application)
-            }
-            NotificationRepository.init(application)
-            NotificationRepository.scanDeviceList(application)
-            _uiState.update { state ->
-                state.copy(deviceList = NotificationRepository.deviceList)
-            }
-        }
-
-        viewModelScope.launch(Dispatchers.IO) {
-            loadInstalledPackages()
-        }
-
         viewModelScope.launch {
             AppRepository.iconUpdates.collect { update ->
                 update?.let { (packageName, _) ->
@@ -85,24 +54,6 @@ class NotificationHistoryViewModel(
                     preloadAppIcons(listOf(packageName))
                 }
             }
-        }
-    }
-
-    fun loadNotifications(device: String) {
-        if (deviceFlow.value == device) {
-            return
-        }
-        deviceFlow.value = device
-        _uiState.update { it.copy(selectedDevice = device, isLoading = true, error = null) }
-
-        viewModelScope.launch(Dispatchers.IO) {
-            NotificationRepository.currentDevice = device
-            NotificationRepository.notifyHistoryChanged(device, application)
-            NotificationRepository.scanDeviceList(application)
-            _uiState.update { state ->
-                state.copy(deviceList = NotificationRepository.deviceList, isLoading = false)
-            }
-            refreshPaging()
         }
     }
 
@@ -116,20 +67,17 @@ class NotificationHistoryViewModel(
         }
     }
 
-    fun deleteNotification(key: String) {
-        val device = _uiState.value.selectedDevice
+    fun deleteEntry(id: Long) {
         viewModelScope.launch(Dispatchers.IO) {
-            NotificationRepository.currentDevice = device
-            NotificationRepository.removeNotification(key, application)
+            repository.deleteSuperIslandHistoryById(id)
             refreshPaging()
         }
     }
 
     fun deleteGroup(packageName: String) {
-        val device = _uiState.value.selectedDevice
         viewModelScope.launch(Dispatchers.IO) {
-            NotificationRepository.currentDevice = device
-            NotificationRepository.removeNotificationsByPackage(packageName, application)
+            val actualPackageName = if (packageName == "(未知应用)") null else packageName
+            repository.deleteSuperIslandHistoryByPackage(actualPackageName)
             _uiState.update { state ->
                 state.copy(expandedGroups = state.expandedGroups - packageName)
             }
@@ -138,10 +86,8 @@ class NotificationHistoryViewModel(
     }
 
     fun clearHistory() {
-        val device = _uiState.value.selectedDevice
         viewModelScope.launch(Dispatchers.IO) {
-            NotificationRepository.currentDevice = device
-            NotificationRepository.clearDeviceHistory(device, application)
+            repository.clearSuperIslandHistory()
             _appIconCache.value = emptyMap()
             _uiState.update { state ->
                 state.copy(expandedGroups = emptySet())
@@ -150,8 +96,12 @@ class NotificationHistoryViewModel(
         }
     }
 
+    suspend fun loadEntryDetail(id: Long): SuperIslandHistoryStoreEntry? {
+        return SuperIslandHistoryStore.loadEntryDetail(application, id)
+    }
+
     fun preloadAppIcons(packageNames: List<String>) {
-        val targets = packageNames.filter { it.isNotBlank() }
+        val targets = packageNames.filter { it.isNotBlank() && it != "(未知应用)" }
         if (targets.isEmpty()) return
 
         val toLoad = synchronized(iconLoading) {
@@ -178,7 +128,7 @@ class NotificationHistoryViewModel(
                 }
             } finally {
                 synchronized(iconLoading) {
-                    iconLoading.removeAll(toLoad.toSet())
+                    iconLoading.removeAll(toLoad)
                 }
             }
         }
@@ -188,15 +138,9 @@ class NotificationHistoryViewModel(
         refreshSignal.value = System.currentTimeMillis()
     }
 
-    private suspend fun loadInstalledPackages() {
-        val cached = AppRepository.getInstalledPackageNames(application)
-        installedPackages.value = cached.ifEmpty {
-            AppRepository.getInstalledPackageNamesAsync(application)
-        }
-    }
-
     private suspend fun getAppNameAndIcon(packageName: String): Pair<String, Bitmap?> {
         var name: String
+        var icon: Bitmap?
         try {
             val pm = application.packageManager
             val appInfo = pm.getApplicationInfo(packageName, 0)
@@ -204,20 +148,20 @@ class NotificationHistoryViewModel(
         } catch (_: Exception) {
             name = packageName
         }
-        val icon: Bitmap? = try {
-            AppRepository.getAppIconWithAutoRequest(application, packageName)
+        try {
+            icon = AppRepository.getAppIconWithAutoRequest(application, packageName)
         } catch (_: Exception) {
-            null
+            icon = null
         }
         return name to icon
     }
 
     class Factory(private val application: Application) : ViewModelProvider.Factory {
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            if (modelClass.isAssignableFrom(NotificationHistoryViewModel::class.java)) {
+            if (modelClass.isAssignableFrom(SuperIslandHistoryViewModel::class.java)) {
                 val repository = DatabaseRepository.getInstance(application)
                 @Suppress("UNCHECKED_CAST")
-                return NotificationHistoryViewModel(application, repository) as T
+                return SuperIslandHistoryViewModel(application, repository) as T
             }
             throw IllegalArgumentException("Unknown ViewModel class")
         }
