+++ base/build.gradle.kts
@@ -1,7 +1,7 @@
 ﻿import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 
 plugins {
-    id("com.android.library")
+    alias(libs.plugins.android.library)
 }
 
 android {
