+++ gradle.properties
@@ -34,7 +34,3 @@ org.gradle.caching=true
 org.gradle.configuration-cache=false
 ## Keep the unsafe flag false unless you have a specific reason to enable it
 org.gradle.unsafe.configuration-cache=false
-android.uniquePackageNames=false
-android.dependency.useConstraints=true
-android.generateSyncIssueWhenLibraryConstraintsAreEnabled=false
-android.r8.strictFullModeForKeepRules=false
